feat/refactor: 좌석 예약 기능 추가 및 리팩토링
- 파사드 패턴 적용

+++ src/main/java/com/hhplu/hhplusconcert/app/interfaces/api/concert/dto/SeatValue.java
@@ -9,8 +9,8 @@ import lombok.Getter;
 @AllArgsConstructor
 public class SeatValue {
     private Long seatId;
-    private int seatNumber;
+    private Long seatNumber;
     private String seatStatus;
-    private int seatPrice;
+    private Long seatPrice;
 }
